Initial traveling trails plus make all the svg embedded in the javascript to improve icon load time
