+++ src/components/journeymap/journeyicons/ClockEndCircle.js
