+++ src/components/journeymap/journeyicons/EndCircle.js
