+++ src/components/journeymap/journeyicons/GavelEndCircle.js
