+++ src/components/journeymap/journeyicons/GavelIcon.js
