+++ src/components/journeymap/journeyicons/StepCompletedCheckbox.js
